move structs into new command folder
