+++ src/command/tag/list.ts
@@ -1,7 +1,7 @@
 import { Message } from 'discord.js';
 import { TagModel } from '../../model/tag';
 import { MessageEmbed } from 'discord.js';
-import { MinehutCommand } from '../../structure/minehutCommand';
+import { MinehutCommand } from '../../structure/command/minehutCommand';
 
 export default class TagListCommand extends MinehutCommand {
 	constructor() {
