+++ src/command/utility/reload.ts
@@ -6,7 +6,7 @@ import {
 } from 'discord-akairo';
 import { messages } from '../../util/messages';
 import { PrefixSupplier } from 'discord-akairo';
-import { MinehutCommand } from '../../structure/minehutCommand';
+import { MinehutCommand } from '../../structure/command/minehutCommand';
 import { PermissionLevel } from '../../util/permission/permissionLevel';
 
 export default class ReloadCommand extends MinehutCommand {
